Se modificó la tabla promoción de la base de datos

+++ Docker/dbo.sql
@@ -279,7 +279,7 @@ GO
 CREATE TABLE [dbo].[Promocion] (
   [id] int  IDENTITY(1,1) NOT NULL,
   [nombre] nvarchar(max) COLLATE Modern_Spanish_CI_AS  NULL,
-  [porcentajeDescuento] decimal(18,2)  NOT NULL,
+  [porcentajeDescuento] INT NOT NULL,
   [cantMaxima] int  NOT NULL,
   [cantMinima] int  NOT NULL
 )
